This code has moved to apollo-kotlin
