--- apollo-execution-runtime/src/commonMain/kotlin/com/apollographql/apollo/execution/LruCache.kt
@@ -1,4 +1,0 @@
-package com.apollographql.apollo.execution
-
-internal typealias Weigher<Key, Value> = (Key, Value?) -> Int
-
