+++ sample-ktor/graphql/schema.graphqls
@@ -3,162 +3,61 @@ schema {
 }
 
 type Query {
-  """
-   Greeting for name
-  """
-  hello(name: String!): String!
+    getPlaylist(id: String): Playlist
 }
 
-type __Schema {
-  description: String
-
-  types: [__Type!]!
-
-  queryType: __Type!
-
-  mutationType: __Type
-
-  subscriptionType: __Type
-
-  directives: [__Directive!]!
+"""
+A playlist
+"""
+type Playlist {
+    """
+    The Spotify ID of the playlist.
+    """
+    id: String
+    """
+    The playlist description. _Only returned for modified, verified playlists, otherwise_ `null`
+    """
+    description: String
+    """
+    The name of the playlist.
+    """
+    name: String
+    """
+    The tracks in the playlist.
+    """
+    tracks: [GetPlaylistTrack]
 }
 
-type __Type {
-  kind: __TypeKind!
-
-  name: String
-
-  description: String
-
-  fields(includeDeprecated: Boolean = false): [__Field!]
-
-  interfaces: [__Type!]
-
-  possibleTypes: [__Type!]
-
-  enumValues(includeDeprecated: Boolean = false): [__EnumValue!]
-
-  inputFields(includeDeprecated: Boolean = false): [__InputValue!]
-
-  ofType: __Type
-
-  specifiedByURL: String
+interface AlbumBase {
+    id: String
+    name: String
 }
 
-enum __TypeKind {
-  SCALAR
-
-  OBJECT
-
-  INTERFACE
-
-  UNION
-
-  ENUM
-
-  INPUT_OBJECT
-
-  LIST
-
-  NON_NULL
+type AlbumObject implements AlbumBase {
+    id: String
+    name: String
+    artists: [SimplifiedArtistObject]
 }
 
-type __Field {
-  name: String!
-
-  description: String
-
-  args(includeDeprecated: Boolean = false): [__InputValue!]!
-
-  type: __Type!
-
-  isDeprecated: Boolean!
-
-  deprecationReason: String
+type QueueObject {
+    currently_playing: CurrentlyPlaying
 }
 
-type __InputValue {
-  name: String!
-
-  description: String
-
-  type: __Type!
-
-  defaultValue: String
-
-  isDeprecated: Boolean!
-
-  deprecationReason: String
+union CurrentlyPlaying = TrackObject | EpisodeObject
+"""
+A track
+"""
+type GetPlaylistTrack {
+    """
+    The Spotify ID for the track.
+    """
+    id: String
+    # more fields
 }
 
-type __EnumValue {
-  name: String!
-
-  description: String
-
-  isDeprecated: Boolean!
-
-  deprecationReason: String
+query GetPlaylist {
+  getPlaylist(id: "42") {
+    id
+    name
+  }
 }
-
-type __Directive {
-  name: String!
-
-  description: String
-
-  locations: [__DirectiveLocation!]!
-
-  args(includeDeprecated: Boolean = false): [__InputValue!]!
-
-  isRepeatable: Boolean!
-}
-
-enum __DirectiveLocation {
-  QUERY
-
-  MUTATION
-
-  SUBSCRIPTION
-
-  FIELD
-
-  FRAGMENT_DEFINITION
-
-  FRAGMENT_SPREAD
-
-  INLINE_FRAGMENT
-
-  VARIABLE_DEFINITION
-
-  SCHEMA
-
-  SCALAR
-
-  OBJECT
-
-  FIELD_DEFINITION
-
-  ARGUMENT_DEFINITION
-
-  INTERFACE
-
-  UNION
-
-  ENUM
-
-  ENUM_VALUE
-
-  INPUT_OBJECT
-
-  INPUT_FIELD_DEFINITION
-}
-
-directive @skip (if: Boolean!) on FIELD|FRAGMENT_SPREAD|INLINE_FRAGMENT
-
-directive @include (if: Boolean!) on FIELD|FRAGMENT_SPREAD|INLINE_FRAGMENT
-
-directive @deprecated (reason: String = "No longer supported") on FIELD_DEFINITION|ARGUMENT_DEFINITION|INPUT_FIELD_DEFINITION|ENUM_VALUE
-
-directive @defer (label: String, if: Boolean! = true) on FRAGMENT_SPREAD|INLINE_FRAGMENT
-
-directive @specifiedBy (url: String!) on SCALAR
